+++ src/controller_device_recv.h
@@ -53,7 +53,7 @@ int device_state_recv_schema_list(device_handle dh, cxobj *xmsg, char *rpcname,
                                   conn_state conn_state);
 int device_state_recv_get_schema(device_handle dh, cxobj *xmsg, char *rpcname,
                                  conn_state conn_state);
-int device_state_recv_ok(device_handle dh, cxobj *xmsg, char *rpcname,
+int device_state_recv_ok(clixon_handle h, device_handle dh, cxobj *xmsg, char *rpcname,
                          conn_state conn_state, cbuf **cberr);
 
 #ifdef __cplusplus